works in configurations
