+++ .gitignore
@@ -14,7 +14,7 @@ build/
 .sts4-cache
 
 ### IntelliJ IDEA ###
-.idea/
+/.idea/
 *.iws
 *.iml
 *.ipr
