+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>ru.netology</groupId>
     <artifactId>servlet</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <version>DI-java-config</version>
     <packaging>war</packaging>
 
     <properties>
@@ -21,6 +21,11 @@
             <artifactId>javax.servlet-api</artifactId>
             <version>4.0.1</version>
             <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+            <version>5.3.12</version>
         </dependency>
         <dependency>
             <groupId>com.google.code.gson</groupId>
