+++ src/main/java/ru/netology/servlet/MainServlet.java
@@ -1,8 +1,11 @@
 package ru.netology.servlet;
 
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import ru.netology.config.JavaConfig;
 import ru.netology.controller.PostController;
 import ru.netology.exception.NotFoundException;
 import ru.netology.repository.PostRepository;
+import ru.netology.repository.PostRepositoryImpl;
 import ru.netology.service.PostService;
 
 import javax.servlet.http.HttpServlet;
@@ -18,14 +21,17 @@ public class MainServlet extends HttpServlet {
     public static final String DELETE = "DELETE";
     private PostController controller;
 
+    final AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(JavaConfig.class);
+
+
     /**
      * Инициализует слои.
      */
     @Override
     public void init() {
-        final var repository = new PostRepository();
-        final var service = new PostService(repository);
-        controller = new PostController(service);
+        final var repository = context.getBean(PostRepositoryImpl.class);
+        final var service = context.getBean(PostService.class);
+        controller = context.getBean(PostController.class);
     }
 
     /**
